+++ rails/app/forms/student/create_task_form.rb
@@ -5,6 +5,7 @@ module Student
     include ActiveModel::Model
     include ActiveModel::Attributes
     include ActiveModel::Validations
+    include UnitIdsValidatable
 
     attr_reader :current_user
 
@@ -47,11 +48,6 @@ module Student
       @goal = current_user.goals.find_by(id: goal_id)
     end
 
-    # フォーム入力を正規化（空文字除去 + integer化）
-    def unit_ids
-      Array(super).compact_blank.map(&:to_i)
-    end
-
     def priority
       value = super
       return if value.blank?
@@ -71,14 +67,6 @@ module Student
       parsed_due_date
     rescue ArgumentError, TypeError
       errors.add(:due_date, 'は正しい日付を入力してください')
-    end
-
-    def unit_ids_must_exist
-      return if unit_ids.blank?
-
-      return unless Unit.where(id: unit_ids).count != unit_ids.size
-
-      errors.add(:unit_ids, 'に不正な値が含まれています')
     end
 
     def goal_must_exist
